refactor(rhiza-utils): split suppression_audit into parse vs. report modules
`.rhiza/utils/suppression_audit.py` was the heaviest module in .rhiza/utils/
(184 statements) and mixed parsing and reporting concerns. Separate them so no
single module dominates the line count, while preserving all public behaviour
and the `make suppression-audit` entry point exactly.

- suppression_parse.py: patterns, the Suppression data model, tokenize-based
  scanning, file collection, density grading, and CVE extraction.
- suppression_report.py: ANSI rendering, the histogram bar, report printing,
  and the pip-audit-backed stale-CVE gate (subprocess invocation).
- suppression_audit.py: thin orchestrator/CLI that wires the two together. It
  inserts its own directory onto sys.path so the sibling modules resolve whether
  run directly, loaded by importlib (tests/fuzz), or frozen by PyInstaller, and
  re-exports the public names (Suppression, scan_file, _nosec_cves, subprocess,
  _active_pip_audit_ids, ...) so existing callers and tests keep working.
- .clusterfuzzlite/build.sh: copy all three modules into tests/fuzz/ so
  PyInstaller bundles the new dependencies of the frozen fuzz binary.

Dogfood-sync: the new/changed .rhiza/utils/ files are mirrored byte-identically
under bundles/core/.rhiza/utils/.

Closes #1310.
